requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
cbor2==5.6.5
cffi==1.17.1
cryptography==45.0.4
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
packaging==25.0
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.1.0
sqlparse==0.5.3
typing_extensions==4.14.0
webauthn==2.5.2
